RuboCop: Fix Naming/ConstantName
Fixes the RuboCop to-do for [Naming/ConstantName](https://docs.rubocop.org/rubocop/0.85/cops_naming.html#namingconstantname).

In this case, RuboCop was incorrectly thinking that `AssertionClass`
should be a constant, likely because of the capitalization used for this
variable. To fix this issue, this variable has been renamed as an
instance variable.

All unit tests:
4543 tests, 72248 assertions, 0 failures, 0 errors, 0 pendings, 0
omissions, 0 notifications
100% passed